+++ src/main/java/uk/gov/hmcts/opal/dto/UpdateDefendantAccountRequest.java
@@ -39,7 +39,7 @@ public class UpdateDefendantAccountRequest implements ToJsonString {
     public static class EnforcementCourtRequest implements ToJsonString {
 
         @JsonProperty("enforcing_court_id")
-        private Integer enforcingCourtId;
+        private Long enforcingCourtId;
 
         @JsonProperty("court_name")
         private String courtName;
@@ -70,7 +70,7 @@ public class UpdateDefendantAccountRequest implements ToJsonString {
         private String enforcementOverrideResultId;
 
         @JsonProperty("enf_override_enforcer_id")
-        private Integer enforcementOverrideEnforcerId;
+        private Long enforcementOverrideEnforcerId;
 
         @JsonProperty("enf_override_tfo_lja_id")
         private Integer enforcementOverrideTfoLjaId;
